fix rubocop linter error and rename the files
