--- myList.rb
@@ -1,21 +1,0 @@
-require_relative 'myEnumerable'
-
-class MyList
-  include MyEnumerable
-
-  def initialize(*items)
-    @list = items
-  end
-
-  def each(&block)
-    @list.each(&block)
-  end
-end
-
-list = MyList.new(1, 2, 3, 4)
-puts list.all? { |e| e < 5 }
-puts list.all? { |e| e > 5 }
-puts list.any? { |e| e == 2 }
-puts list.any? { |e| e == 5 }
-puts list.filter { |e| e.even? }.to_s
-puts list.filter { |e| e.odd? }.to_s
